Renaming, cleanup and small fixes
The close button checks if there's no tabs left.
openFile() method checks if the selected file is opened before.
Renaming of all components are completed.
Added workspace option to the preferences and Properties.java
Moved the components Find & Replace frame into a panel.
Replaced backslashes with slashes on FileConfigurer.java
Cleared and renamed the old folders in the repository.