Integration of runutils
FileConfigurer is now handling also the build.xml files by using a new
class BuildConfigurer.
Compile button now puts the class files into a classes folder that is
generated next to the source file.
Run button searches for a classes folder next to the source file and
runs the class file in it if found.