New features and many bugfixes
Popup menu operations are completed.
Compile and run button is enabled for projects as well.
File drop bug about new tabs is fixed.
Some of the bugs are fixed about project system.
Compiling and generating jar is now working properly with packages.
Different type of checks are added to various operations.
Readability improvements are done on the code.
Some of the javadocs are improved, some new ones are added.